LayoutTests/platform/ios-sim-deprecated/svg/animations/change-target-while-animating-SVG-property-expected.txt

SVG 1.1 dynamic animation tests

This changes the target of an animation while its running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect1.width.animVal.value is 150
PASS rect1.width.baseVal.value is 150
PASS rect2.width.animVal.value is 150
PASS rect2.width.baseVal.value is 150
PASS rect1.width.animVal.value is 100
PASS rect1.width.baseVal.value is 150
PASS rect2.width.animVal.value is 150
PASS rect2.width.baseVal.value is 150
PASS rect1.width.animVal.value is 150
PASS rect1.width.baseVal.value is 150
PASS rect2.width.animVal.value is 100
PASS rect2.width.baseVal.value is 150
PASS rect1.width.animVal.value is 150
PASS rect1.width.baseVal.value is 150
PASS rect2.width.animVal.value is 50
PASS rect2.width.baseVal.value is 150
PASS rect1.width.animVal.value is 150
PASS rect1.width.baseVal.value is 150
PASS rect2.width.animVal.value is 50
PASS rect2.width.baseVal.value is 150
PASS rect1.width.animVal.value is 150
PASS rect1.width.baseVal.value is 150
PASS rect2.width.animVal.value is 50
PASS rect2.width.baseVal.value is 150
PASS successfullyParsed is true

TEST COMPLETE
